Bump, move components to shared
