+++ shared/components/TileButton/TileButton.js
@@ -1,8 +1,17 @@
 import React from 'react';
 
+import PropTypes from 'prop-types';
 import classNames from 'classnames';
 
 import './TileButton.scss';
+
+TileButton.propTypes = {
+  title: PropTypes.string.isRequired,
+  description: PropTypes.string,
+  icon: PropTypes.node.isRequired,
+  isActive: PropTypes.bool,
+  handleClick: PropTypes.func.isRequired,
+};
 
 export function TileButton({
   title,
