+++ shared/components/VerticalTabs/VerticalTabs.scss
