+++ shared/index.js
@@ -35,6 +35,8 @@ export * from './components/ResourceDetails/ResourceDetails';
 export * from './components/ReadableCreationTimestamp/ReadableCreationTimestamp';
 export * from './components/KeyValueForm/KeyValueForm';
 export * from './components/MainFrameRedirection/MainFrameRedirection';
+export * from './components/TileButton/TileButton';
+export * from './components/VerticalTabs/VerticalTabs';
 
 export * from './forms';
 export * from './hooks';
